Added float_expression evaluator, added U gate, swithed back from u1,u2,u3 to U and cx to CX, Added circuit printing

+++ include/Parser/float_expr_ast.hpp
@@ -14,7 +14,7 @@ namespace Parser {
         struct t_float_expression;
         class OperandPrinterVisitor;
 
-        typedef ::boost::spirit::x3::variant<float, std::string> t_float;
+        typedef ::boost::spirit::x3::variant<double, std::string> t_float;
 
         struct t_float_expr_nil {  
         };
